+++ feature/mybookdetail/src/main/java/app/kaito_dogi/mybrary/feature/mybookdetail/MyBookDetailViewModel.kt
@@ -1,4 +1,4 @@
-package app.kaito_dogi.feature.mybookdetail
+package app.kaito_dogi.mybrary.feature.mybookdetail
 
 import androidx.lifecycle.ViewModel
 import dagger.hilt.android.lifecycle.HiltViewModel
