+++ feature/searchbook/src/main/java/app/kaito_dogi/mybrary/feature/searchbook/SearchBookViewModel.kt
@@ -1,4 +1,4 @@
-package app.kaito_dogi.feature.searchbook
+package app.kaito_dogi.mybrary.feature.searchbook
 
 import androidx.lifecycle.ViewModel
 import dagger.hilt.android.lifecycle.HiltViewModel
